pyproject.toml
[project]
name = "dashboard-01"
version = "0.1.0"
description = "Dashboard project with MkDocs and data visualization"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocs-jupyter>=0.24",
    "mkdocs-macros-plugin>=1.0",
    "mkdocs-gen-files>=0.5",
    "mkdocs-minify-plugin>=0.7",
    "pyyaml>=6.0",
    "pandas>=2.2",
    "matplotlib>=3.8",
    "openpyxl>=3.1"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
